+++ module-apps/application-meditation/widgets/MeditationModel.hpp
@@ -22,7 +22,7 @@ namespace gui
         {}
 
         [[nodiscard]] auto requestRecordsCount() -> unsigned int final;
-        [[nodiscard]] auto getMinimalItemHeight() const -> unsigned int final;
+        [[nodiscard]] auto getMinimalItemSpaceRequired() const -> unsigned int final;
         auto getItem(gui::Order order) -> gui::ListItem * final;
         void requestRecords(const uint32_t offset, const uint32_t limit) final;
     };
