+++ module-apps/application-messages/models/SMSThreadModel.hpp
@@ -32,6 +32,6 @@ class SMSThreadModel : public app::DatabaseModel<SMSRecord>,
     unsigned int requestRecordsCount() override;
     bool updateRecords(std::vector<SMSRecord> records) override;
     void requestRecords(uint32_t offset, uint32_t limit) override;
-    unsigned int getMinimalItemHeight() const override;
+    unsigned int getMinimalItemSpaceRequired() const override;
     gui::ListItem *getItem(gui::Order order) override;
 };
