+++ module-apps/application-messages/models/ThreadsModel.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2020, Mudita Sp. z.o.o. All rights reserved.
+// Copyright (c) 2017-2021, Mudita Sp. z.o.o. All rights reserved.
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
 #pragma once
@@ -12,7 +12,7 @@ class ThreadsModel : public BaseThreadsRecordModel, public app::AsyncCallbackRec
     explicit ThreadsModel(app::Application *app);
 
     void requestRecords(uint32_t offset, uint32_t limit) override;
-    [[nodiscard]] auto getMinimalItemHeight() const -> unsigned int override;
+    [[nodiscard]] auto getMinimalItemSpaceRequired() const -> unsigned int override;
     [[nodiscard]] auto getItem(gui::Order order) -> gui::ListItem * override;
 
     auto handleQueryResponse(db::QueryResult *queryResult) -> bool;
