+++ module-apps/application-phonebook/models/NewContactModel.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2020, Mudita Sp. z.o.o. All rights reserved.
+// Copyright (c) 2017-2021, Mudita Sp. z.o.o. All rights reserved.
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
 #pragma once
@@ -24,7 +24,7 @@ class NewContactModel : public app::InternalModel<gui::ContactListItem *>, publi
 
     [[nodiscard]] auto requestRecordsCount() -> unsigned int override;
 
-    [[nodiscard]] auto getMinimalItemHeight() const -> unsigned int override;
+    [[nodiscard]] auto getMinimalItemSpaceRequired() const -> unsigned int override;
 
     auto getItem(gui::Order order) -> gui::ListItem * override;
