+++ module-apps/application-settings/models/display-keypad/CategoriesModel.hpp
@@ -17,7 +17,7 @@ namespace Quotes
       public:
         explicit CategoriesModel(app::Application *app);
         [[nodiscard]] auto requestRecordsCount() -> unsigned int final;
-        [[nodiscard]] auto getMinimalItemHeight() const -> unsigned int final;
+        [[nodiscard]] auto getMinimalItemSpaceRequired() const -> unsigned int final;
         auto getItem(gui::Order order) -> gui::ListItem * final;
         void requestRecords(const uint32_t offset, const uint32_t limit) final;
         bool updateRecords(std::vector<CategoryRecord> records);
