+++ module-apps/application-settings/models/display-keypad/QuotesModel.hpp
@@ -23,7 +23,7 @@ namespace Quotes
         explicit QuotesModel(app::Application *application);
 
         [[nodiscard]] auto requestRecordsCount() -> unsigned int final;
-        [[nodiscard]] auto getMinimalItemHeight() const -> unsigned int final;
+        [[nodiscard]] auto getMinimalItemSpaceRequired() const -> unsigned int final;
 
         auto getItem(gui::Order order) -> gui::ListItem * final;
         void requestRecords(const uint32_t offset, const uint32_t limit) final;
