+++ module-apps/apps-common/options/OptionsModel.hpp
@@ -18,7 +18,7 @@ class OptionsModel : public app::InternalModel<gui::ListItem *>, public gui::Lis
     void createData(std::list<gui::Option> &optionList);
 
     [[nodiscard]] auto requestRecordsCount() -> unsigned int override;
-    [[nodiscard]] auto getMinimalItemHeight() const -> unsigned int override;
+    [[nodiscard]] auto getMinimalItemSpaceRequired() const -> unsigned int override;
 
     auto getItem(gui::Order order) -> gui::ListItem * override;
