+++ module-gui/gui/widgets/CMakeLists.txt
@@ -14,6 +14,7 @@ target_sources( ${PROJECT_NAME}
         "${CMAKE_CURRENT_LIST_DIR}/ListItem.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ListItemWithDescription.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ListView.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/ListViewEngine.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/Margins.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/Navigation.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/ProgressBar.cpp"
@@ -72,6 +73,7 @@ target_sources( ${PROJECT_NAME}
         "${CMAKE_CURRENT_LIST_DIR}/ListItemWithDescription.hpp"
         "${CMAKE_CURRENT_LIST_DIR}/ListItemProvider.hpp"
         "${CMAKE_CURRENT_LIST_DIR}/ListView.hpp"
+        "${CMAKE_CURRENT_LIST_DIR}/ListViewEngine.hpp"
         "${CMAKE_CURRENT_LIST_DIR}/Margins.hpp"
         "${CMAKE_CURRENT_LIST_DIR}/Navigation.hpp"
         "${CMAKE_CURRENT_LIST_DIR}/ProgressBar.hpp"
