+++ module-gui/gui/widgets/ListView.cpp
@@ -8,7 +8,6 @@
 
 namespace gui
 {
-
     ListViewScroll::ListViewScroll(
         Item *parent, unsigned int x, unsigned int y, unsigned int w, unsigned int h, listview::ScrollBarType type)
         : Rect{parent, x, y, w, h}, type(type)
@@ -23,10 +22,10 @@ namespace gui
     void ListViewScroll::updateProportional(const ListViewScrollUpdateData &data)
     {
         double scrollStep =
-            static_cast<double>((parent->widgetArea.h - data.topMargin)) / static_cast<double>(data.elementsCount);
+            static_cast<double>((parent->widgetArea.h - topMargin)) / static_cast<double>(data.elementsCount);
 
         auto scrollH = scrollStep * data.listPageSize;
-        auto scrollY = scrollStep * data.startIndex > 0 ? scrollStep * data.startIndex : data.topMargin;
+        auto scrollY = scrollStep * data.startIndex > 0 ? scrollStep * data.startIndex : topMargin;
 
         setArea(BoundingBox(
             parent->widgetArea.w - style::listview::scroll::margin, scrollY, style::listview::scroll::w, scrollH));
@@ -41,8 +40,8 @@ namespace gui
 
         currentPage = data.startIndex / elementsOnPage;
 
-        auto scrollH = (parent->widgetArea.h - data.topMargin) / pagesCount;
-        auto scrollY = scrollH * currentPage > 0 ? scrollH * currentPage : data.topMargin;
+        auto scrollH = (parent->widgetArea.h - topMargin) / pagesCount;
+        auto scrollY = scrollH * currentPage > 0 ? scrollH * currentPage : topMargin;
 
         setArea(BoundingBox(
             parent->widgetArea.w - style::listview::scroll::margin, scrollY, style::listview::scroll::w, scrollH));
@@ -72,19 +71,17 @@ namespace gui
         storedStartIndex = data.startIndex;
 
         auto scrollH = parent->widgetArea.h / pagesCount;
-        auto scrollY = currentPage * scrollH > 0 ? currentPage * scrollH : data.topMargin;
+        auto scrollY = currentPage * scrollH > 0 ? currentPage * scrollH : topMargin;
 
         setArea(BoundingBox(
             parent->widgetArea.w - style::listview::scroll::margin, scrollY, style::listview::scroll::w, scrollH));
     }
 
-    void ListViewScroll::updateStartConditions(const unsigned int index,
-                                               const unsigned int page,
-                                               const unsigned int count)
+    void ListViewScroll::updateStartConditions(const ListViewScrollSetupData &data)
     {
-        storedStartIndex = index;
-        currentPage      = page;
-        pagesCount       = count;
+        storedStartIndex = data.storedStartIndex;
+        currentPage      = data.currentPage;
+        pagesCount       = data.pagesCount;
     }
 
     bool ListViewScroll::shouldShowScroll(unsigned int currentPageSize, unsigned int elementsCount)
@@ -116,18 +113,9 @@ namespace gui
         else
             setVisible(false);
     }
-
-    ListView::ListView()
+    void ListViewScroll::setTopMargin(int _topMargin)
     {
-        body = new VBox{this, 0, 0, 0, 0};
-
-        scroll = new ListViewScroll(this,
-                                    style::listview::scroll::x,
-                                    style::listview::scroll::y,
-                                    style::listview::scroll::w,
-                                    style::listview::scroll::h,
-                                    listview::ScrollBarType::None);
-        type   = gui::ItemType::LIST;
+        topMargin = _topMargin;
     }
 
     ListView::ListView(Item *parent,
@@ -137,9 +125,8 @@ namespace gui
                        unsigned int h,
                        std::shared_ptr<ListItemProvider> prov,
                        listview::ScrollBarType scrollBarType)
-        : Rect{parent, x, y, w, h}
+        : Rect{parent, x, y, w, h}, ListViewEngine(prov)
     {
-
         this->setBorderColor(ColorNoColor);
 
         body = new VBox{this, 0, 0, w, h};
@@ -176,6 +163,12 @@ namespace gui
                 recalculateOnBoxRequestedResize();
         };
 
+        checkFullRenderRequirementCallback = [this]() {
+            if (scroll && scroll->type == listview::ScrollBarType::PreRendered) {
+                requestFullListRender = true;
+            }
+        };
+
         if (scrollBarType != listview::ScrollBarType::None) {
             scroll = new ListViewScroll(this,
                                         style::listview::scroll::x,
@@ -185,171 +178,50 @@ namespace gui
                                         scrollBarType);
         }
 
-        setProvider(std::move(prov));
+        applyScrollCallbacks();
 
         type = gui::ItemType::LIST;
     }
 
-    ListView::~ListView()
+    void ListView::applyScrollCallbacks()
     {
-        clear();
+        setupScrollCallback = [this](ListViewScrollSetupData data) {
+            if (scroll) {
+                scroll->updateStartConditions(data);
+            }
+        };
+
+        updateScrollCallback = [this](ListViewScrollUpdateData data) {
+            if (scroll) {
+                scroll->update(data);
+            }
+        };
+
+        resizeScrollCallback = [this]() {
+            if (scroll && scroll->shouldShowScroll(currentPageSize, elementsCount)) {
+                body->setSize(style::listview::item_width_with_scroll, body->getHeight());
+            }
+            else {
+                body->setSize(style::window::default_body_width, body->getHeight());
+            }
+        };
     }
 
-    void ListView::setElementsCount(unsigned int count)
+    void ListView::setFocus()
     {
-        if (elementsCount != count) {
-            elementsCount = count;
-            onElementsCountChanged();
+        if (!focus) {
+            return;
         }
-    }
+        setFocusItem(body);
 
-    void ListView::setBoundaries(Boundaries value)
-    {
-        boundaries = value;
+        ListViewEngine::setFocus();
     }
 
     void ListView::setScrollTopMargin(int value)
     {
-        scrollTopMargin = value;
-    }
-
-    void ListView::setOrientation(listview::Orientation value)
-    {
-        orientation = value;
-
-        if (orientation == listview::Orientation::TopBottom) {
-            body->setAlignment(Alignment::Vertical::Top);
+        if (scroll) {
+            scroll->setTopMargin(value);
         }
-        else {
-            body->setAlignment(Alignment::Vertical::Bottom);
-        }
-    }
-
-    void ListView::setFocusOnElement(unsigned int elementNumber)
-    {
-        body->setFocusOnElement(elementNumber);
-    }
-
-    void ListView::setProvider(std::shared_ptr<ListItemProvider> prov)
-    {
-        if (prov != nullptr) {
-            provider       = prov;
-            provider->list = this;
-        }
-    }
-
-    bool ListView::isEmpty() const noexcept
-    {
-        return elementsCount == 0;
-    }
-
-    void ListView::rebuildList(listview::RebuildType rebuildType, unsigned int dataOffset, bool forceRebuild)
-    {
-        if (pageLoaded || forceRebuild) {
-
-            setElementsCount(provider->requestRecordsCount());
-
-            setup(rebuildType, dataOffset);
-
-            // If deletion operation caused last page to be removed request previous one.
-            if ((startIndex != 0 && startIndex == elementsCount)) {
-                requestPreviousPage();
-            }
-            else {
-                provider->requestRecords(startIndex, calculateLimit());
-            }
-        }
-        else {
-            rebuildRequests.push_front({rebuildType, dataOffset});
-        }
-    }
-
-    void ListView::reSendLastRebuildRequest()
-    {
-        rebuildList(lastRebuildRequest.first, lastRebuildRequest.second, true);
-    }
-
-    void ListView::prepareFullRebuild()
-    {
-        setStartIndex();
-        storedFocusIndex = listview::nPos;
-    }
-
-    void ListView::prepareOnOffsetRebuild(unsigned int dataOffset)
-    {
-        if (dataOffset < elementsCount) {
-            startIndex       = dataOffset;
-            storedFocusIndex = listview::nPos;
-        }
-        else {
-            LOG_ERROR("Requested rebuild on index greater than elements count");
-        }
-    }
-
-    void ListView::prepareInPlaceRebuild()
-    {
-        if (!body->empty()) {
-            storedFocusIndex = getFocusItemIndex();
-        }
-    }
-
-    void ListView::prepareOnPageElementRebuild(unsigned int dataOffset)
-    {
-        startIndex       = (dataOffset / calculateMaxItemsOnPage()) * calculateMaxItemsOnPage();
-        storedFocusIndex = dataOffset % calculateMaxItemsOnPage();
-    }
-
-    void ListView::setup(listview::RebuildType rebuildType, unsigned int dataOffset)
-    {
-        switch (rebuildType) {
-        case listview::RebuildType::Full:
-            prepareFullRebuild();
-            break;
-        case listview::RebuildType::OnOffset:
-            prepareOnOffsetRebuild(dataOffset);
-            break;
-        case listview::RebuildType::InPlace:
-            prepareInPlaceRebuild();
-            break;
-        case listview::RebuildType::OnPageElement:
-            prepareOnPageElementRebuild(dataOffset);
-            break;
-        }
-
-        if (prepareRebuildCallback) {
-            prepareRebuildCallback();
-            setElementsCount(provider->requestRecordsCount());
-        }
-
-        lastRebuildRequest = {rebuildType, dataOffset};
-
-        body->setReverseOrder(false);
-        direction = listview::Direction::Bottom;
-    }
-
-    void ListView::onClose()
-    {
-        if (!body->empty()) {
-            storedFocusIndex = getFocusItemIndex();
-        }
-        clear();
-    }
-
-    unsigned int ListView::getFocusItemIndex()
-    {
-        auto index = body->getFocusItemIndex();
-
-        if (direction == listview::Direction::Top) {
-            int position = currentPageSize - 1 - index;
-            index        = std::abs(position);
-        }
-
-        return index;
-    }
-
-    std::shared_ptr<ListItemProvider> ListView::getProvider()
-    {
-        return provider;
     }
 
     void ListView::setAlignment(const Alignment &value)
@@ -359,278 +231,18 @@ namespace gui
         }
     }
 
-    void ListView::reset()
-    {
-        clear();
-        setStartIndex();
-        body->setReverseOrder(false);
-        direction = listview::Direction::Bottom;
-    }
-
-    void ListView::clear()
-    {
-        body->setFocusItem(nullptr);
-
-        while (auto el = body->children.back()) {
-
-            if (el->type == ItemType::LIST_ITEM) {
-
-                if (!dynamic_cast<ListItem *>(el)->deleteByList) {
-                    body->removeWidget(el);
-                }
-                else {
-                    body->erase(el);
-                }
-            }
-            else
-                body->erase(el);
-        }
-    }
-
-    void ListView::refresh()
-    {
-        if (provider == nullptr) {
-            LOG_ERROR("ListView Data provider not exist");
-            return;
-        }
-
-        clear();
-
-        addItemsOnPage();
-
-        setFocus();
-        if (scroll) {
-            scroll->update(ListViewScrollUpdateData{startIndex,
-                                                    currentPageSize,
-                                                    elementsCount,
-                                                    provider->getMinimalItemHeight(),
-                                                    direction,
-                                                    boundaries,
-                                                    scrollTopMargin});
-        }
-        resizeWithScroll();
-        pageLoaded = true;
-
-        // Check if there were queued rebuild Requests - if so rebuild list again.
-        if (!rebuildRequests.empty()) {
-            rebuildList(rebuildRequests.back().first, rebuildRequests.back().second);
-            rebuildRequests.pop_back();
-        }
-
-        fillFirstPage();
-    }
-
-    void ListView::onProviderDataUpdate()
-    {
-        if (!renderFullList()) {
-            return;
-        }
-
-        refresh();
-    }
-
-    Order ListView::getOrderFromDirection() const noexcept
-    {
-        if (direction == listview::Direction::Bottom)
-            return Order::Next;
-
-        return Order::Previous;
-    }
-
-    Order ListView::getOppositeOrderFromDirection() const noexcept
-    {
-        if (direction == listview::Direction::Bottom)
-            return Order::Previous;
-
-        return Order::Next;
-    }
-
-    void ListView::setStartIndex()
-    {
-        if (orientation == listview::Orientation::TopBottom) {
-            startIndex = 0;
-        }
-        else {
-            startIndex = elementsCount;
-        }
-    }
-
-    void ListView::recalculateStartIndex()
-    {
-        if (direction == listview::Direction::Top) {
-            startIndex = startIndex < currentPageSize ? 0 : startIndex - currentPageSize;
-        }
-    }
-
-    void ListView::fillFirstPage()
-    {
-        // Check if first page is filled with items. If not reload page to be filed with items. Check for both
-        // Orientations.
-        if (orientation == listview::Orientation::TopBottom && direction == listview::Direction::Top &&
-            startIndex == 0) {
-            if (body->getSizeLeft() > provider->getMinimalItemHeight()) {
-                focusOnLastItem = true;
-
-                checkFullRenderRequirement();
-                rebuildList();
-            }
-        }
-
-        if (orientation == listview::Orientation::BottomTop && direction == listview::Direction::Bottom &&
-            startIndex + currentPageSize == elementsCount) {
-            if (body->getSizeLeft() > provider->getMinimalItemHeight()) {
-                focusOnLastItem = true;
-
-                checkFullRenderRequirement();
-                rebuildList();
-            }
-        }
-    }
-
-    void ListView::resizeWithScroll()
-    {
-        if (scroll && scroll->shouldShowScroll(currentPageSize, elementsCount)) {
-            body->setSize(style::listview::item_width_with_scroll, body->getHeight());
-        }
-        else {
-            body->setSize(style::window::default_body_width, body->getHeight());
-        }
-    }
-
-    void ListView::addItemsOnPage()
-    {
-        currentPageSize = 0;
-
-        ListItem *item = nullptr;
-
-        while ((item = provider->getItem(getOrderFromDirection())) != nullptr) {
-
-            body->addWidget(item);
-
-            if (!item->visible) {
-                // In case model is tracking internal indexes -> undo last get.
-                if (requestFullListRender) {
-                    provider->getItem(getOppositeOrderFromDirection());
-                }
-                break;
-            }
-
-            currentPageSize++;
-        }
-
-        recalculateStartIndex();
-    }
-
-    void ListView::checkFullRenderRequirement()
-    {
-        if (scroll && scroll->type == listview::ScrollBarType::PreRendered) {
-            requestFullListRender = true;
-        }
-    }
-
-    bool ListView::renderFullList()
-    {
-        if (!requestFullListRender) {
-            return true;
-        }
-
-        if (elementsCount != 0 && !requestCompleteData) {
-            requestCompleteData = true;
-            provider->requestRecords(0, elementsCount);
-            return false;
-        }
-
-        if (requestCompleteData) {
-
-            auto page           = 0;
-            auto pageStartIndex = 0;
-
-            clear();
-
-            while (true) {
-
-                addItemsOnPage();
-
-                if (currentPageSize == 0) {
-                    break;
-                }
-
-                if (currentPageSize + pageStartIndex == elementsCount) {
-                    break;
-                }
-
-                page += 1;
-                pageStartIndex += currentPageSize;
-
-                clear();
-            }
-
-            clear();
-            requestCompleteData   = false;
-            requestFullListRender = false;
-
-            if (lastRebuildRequest.first == listview::RebuildType::Full) {
-                if (orientation == listview::Orientation::TopBottom) {
-                    scroll->updateStartConditions(startIndex, 0, page + 1);
-                }
-                else {
-                    scroll->updateStartConditions(startIndex, page, page + 1);
-                }
-            }
-
-            reSendLastRebuildRequest();
-            return false;
-        }
-
-        return true;
-    } // namespace gui
-
-    void ListView::setFocus()
-    {
-        if (!focus) {
-            return;
-        }
-
-        setFocusItem(body);
-
-        if (storedFocusIndex != listview::nPos) {
-            if (!body->setFocusOnElement(storedFocusIndex)) {
-                body->setFocusOnLastElement();
-            }
-        }
-
-        if (focusOnLastItem) {
-            body->setFocusOnLastElement();
-            focusOnLastItem = false;
-        }
-    };
-
-    void ListView::onElementsCountChanged()
-    {
-        if (isEmpty()) {
-            if (emptyListCallback) {
-                emptyListCallback();
-            }
-        }
-        else if (notEmptyListCallback) {
-            notEmptyListCallback();
-        }
-
-        checkFullRenderRequirement();
-    }
-
     bool ListView::onDimensionChanged(const BoundingBox &oldDim, const BoundingBox &newDim)
     {
         Rect::onDimensionChanged(oldDim, newDim);
         body->setSize(body->getWidth(), newDim.h);
-        if (scroll) {
-            scroll->update(ListViewScrollUpdateData{startIndex,
-                                                    currentPageSize,
-                                                    elementsCount,
-                                                    provider->getMinimalItemHeight(),
-                                                    direction,
-                                                    boundaries,
-                                                    scrollTopMargin});
+
+        if (updateScrollCallback) {
+            updateScrollCallback(ListViewScrollUpdateData{startIndex,
+                                                          currentPageSize,
+                                                          elementsCount,
+                                                          provider->getMinimalItemSpaceRequired(),
+                                                          direction,
+                                                          boundaries});
         }
 
         return true;
@@ -638,108 +250,11 @@ namespace gui
 
     auto ListView::handleRequestResize([[maybe_unused]] const Item *child, Length request_w, Length request_h) -> Size
     {
-
         return Size(request_w, request_h);
-    }
-
-    void ListView::recalculateOnBoxRequestedResize()
-    {
-        if (currentPageSize != body->getVisibleChildrenCount()) {
-
-            unsigned int diff = currentPageSize < body->getVisibleChildrenCount()
-                                    ? 0
-                                    : currentPageSize - body->getVisibleChildrenCount();
-            currentPageSize = body->getVisibleChildrenCount();
-
-            if (direction == listview::Direction::Top) {
-                startIndex += diff;
-            }
-            else {
-                startIndex = startIndex < diff ? 0 : startIndex - diff;
-            }
-
-            checkFullRenderRequirement();
-            rebuildList();
-        }
     }
 
     bool ListView::onInput(const InputEvent &inputEvent)
     {
         return body->onInput(inputEvent);
-    }
-
-    unsigned int ListView::calculateMaxItemsOnPage()
-    {
-        assert(provider->getMinimalItemHeight() != 0);
-        auto count = widgetArea.h / provider->getMinimalItemHeight();
-
-        return count;
-    }
-
-    unsigned int ListView::calculateLimit(listview::Direction value)
-    {
-        auto minLimit =
-            (2 * currentPageSize > calculateMaxItemsOnPage() ? 2 * currentPageSize : calculateMaxItemsOnPage());
-        if (value == listview::Direction::Bottom)
-            return (minLimit + startIndex <= elementsCount ? minLimit : elementsCount - startIndex);
-        else
-            return minLimit < startIndex ? minLimit : startIndex;
-    }
-
-    bool ListView::requestNextPage()
-    {
-        if (startIndex + currentPageSize >= elementsCount && boundaries == Boundaries::Continuous) {
-
-            startIndex = 0;
-        }
-        else if (startIndex + currentPageSize >= elementsCount && boundaries == Boundaries::Fixed) {
-
-            return false;
-        }
-        else {
-
-            startIndex = startIndex <= elementsCount - currentPageSize ? startIndex + currentPageSize
-                                                                       : elementsCount - (elementsCount - startIndex);
-        }
-
-        direction = listview::Direction::Bottom;
-        body->setReverseOrder(false);
-        pageLoaded       = false;
-        storedFocusIndex = listview::nPos;
-        provider->requestRecords(startIndex, calculateLimit());
-
-        return true;
-    }
-
-    bool ListView::requestPreviousPage()
-    {
-        auto topFetchIndex = 0;
-        auto limit         = 0;
-
-        if (startIndex == 0 && boundaries == Boundaries::Continuous) {
-
-            startIndex    = elementsCount;
-            topFetchIndex = elementsCount - calculateLimit(listview::Direction::Top);
-            limit         = calculateLimit(listview::Direction::Top);
-        }
-        else if (startIndex == 0 && boundaries == Boundaries::Fixed) {
-
-            return false;
-        }
-        else {
-
-            limit         = calculateLimit(listview::Direction::Top);
-            topFetchIndex = startIndex < calculateLimit(listview::Direction::Top)
-                                ? 0
-                                : startIndex - calculateLimit(listview::Direction::Top);
-        }
-
-        direction = listview::Direction::Top;
-        body->setReverseOrder(true);
-        pageLoaded       = false;
-        storedFocusIndex = listview::nPos;
-        provider->requestRecords(topFetchIndex, limit);
-
-        return true;
     }
 } /* namespace gui */
