+++ module-gui/test/mock/TestListViewProvider.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2020, Mudita Sp. z.o.o. All rights reserved.
+// Copyright (c) 2017-2021, Mudita Sp. z.o.o. All rights reserved.
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
 #include "TestListViewProvider.hpp"
@@ -34,7 +34,7 @@ namespace gui
         return testItemCount;
     }
 
-    unsigned int TestListViewProvider::getMinimalItemHeight() const
+    unsigned int TestListViewProvider::getMinimalItemSpaceRequired() const
     {
 
         return testItemMinimalHeight;
